+++ dev/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.5.0](https://github.com/google/adk-js/compare/devtools-v1.4.0...devtools-v1.5.0) (2026-07-29)
+
+
+### Bug Fixes
+
+* **security:** require authentication when mounting A2A server ([#529](https://github.com/google/adk-js/issues/529)) ([7f39b6c](https://github.com/google/adk-js/commit/7f39b6c330bcadfedc35c29d67d103e7d0f6bb2b))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @google/adk bumped from ^1.4.0 to ^1.5.0
+
 ## [1.4.0](https://github.com/google/adk-js/compare/devtools-v1.3.0...devtools-v1.4.0) (2026-07-20)